pyproject.toml
[project]
name = "kafka-svc"
version = "0.1.0"
description = "Kafka Svc microservice"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
  "alembic==1.15.2",
  "celery==5.4.0",
  "ruff==0.11.2",
  "setuptools==78.1.0",
  "sqlalchemy==2.0.40",
  "sqlalchemy-utils==0.41.2",
  "psycopg2-binary==2.9.10",
  "psycopg==3.2.6",
  "redis==5.2.1",
  "grpcio==1.72.0",
  "grpcio-tools==1.72.0",
  "grpcio-reflection==1.72.0",
  "pydantic==2.11.3",
  "pydantic-settings==2.9.1",
  "asyncpg==0.30.0",
  "pytest-asyncio==0.26.0",
  "aiokafka==0.12.0",
  "PyYAML==6.0.2",
  "grpcio==1.72.0",
  "grpcio-tools==1.72.0",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
extend-select = ["I", "SIM"]
